Remove Send and Sync autotraits from the lock guard
Having Send on the lock guard is unsound.
Having Sync on the lock guard is unsound, unless T is Sync.

Normally this should use negative trait impls to get rid of Send and Sync.
However, negative trait impls are not fully stabilized, yet.
Use a phantom pointer instead. That's a bit overly strict in that it also removes Sync, if T is Sync.